ajout de la gestion de l’affichage des interactions problème avec click sur play après une mis en pause programmée boutons d’interactions ne disparaisse pas si je clique sur play

+++ SmartVideoUrl/Entities.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 struct InteractionBtn {
-    let label: String, goto: Float?
-    init(label: String, goto: Float?) {
+    let label: String, goto: Double?
+    init(label: String, goto: Double?) {
         self.label = label
         self.goto = goto
     }
